Remove bad workarounds to solve the React event pooling problem on tap
The solution is to fire the tap event always synchronously (already fixed).
So the React event pooling can behave as usual.

Removed the attempt to assign an empty function to preventDefault, because it does not make sense to assign a function to a SyntheticEvent that has been put back in the pool. It should be the role of React to be defensive agaiin using preventDefault on events that are already put back in the pool. (See https://github.com/facebook/react/issues/3657#issuecomment-125681594 )